+++ Aras.VS.MethodPlugin.Shared/Commands/UpdateMethodCmd.cs
@@ -81,7 +81,7 @@ namespace Aras.VS.MethodPlugin.Commands
 			MethodInfo methodInformation = projectConfigurationManager.CurrentProjectConfiguraiton.MethodInfos.FirstOrDefault(m => m.MethodName == selectedMethodName);
 			if (methodInformation == null)
 			{
-				throw new Exception();
+				throw new Exception(this.messageManager.GetMessage("ConfigurationsForTheMethodNotFound", selectedMethodName));
 			}
 
 			TemplateLoader templateLoader = new TemplateLoader();